requirements.txt
streamlit==1.29.0
pandas==2.2.0
numpy==1.26.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
plotly==5.17.0
sqlalchemy==2.0.23
openpyxl==3.1.2
